Added arm64 support to UtilsDtos Class

+++ Owncloud iOs Client/Utils/UtilsDtos.m
@@ -39,7 +39,7 @@
     urlWithoutAddress = [filePath substringFromIndex:[urlWithoutAddress length]];
     
     //NSString *newLocalFolder= [[NSHomeDirectory() stringByAppendingPathComponent:@"Library/Caches"] stringByAppendingPathComponent:[NSString stringWithFormat:@"%d", mUser.idUser]];
-    NSString *newLocalFolder= [[UtilsUrls getOwnCloudFilePath] stringByAppendingPathComponent:[NSString stringWithFormat:@"%d", mUser.idUser]];
+    NSString *newLocalFolder= [[UtilsUrls getOwnCloudFilePath] stringByAppendingPathComponent:[NSString stringWithFormat:@"%ld", (long)mUser.idUser]];
